# docs/myblog-api-openapi.yaml
openapi: 3.0.3
info:
  title: Myblog API
  description: 个人博客系统后端API接口文档
  version: 3.0.0
  contact:
    name: 星辰远征客
    email: 1850833838@qq.com
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
...
paths:
...
  /auth/login:
    post:
      tags:
        - 认证
      summary: 用户登录
      description: 用户登录接口，限流5次/5分钟
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: 147258369Thy@
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
  /auth/logout:
    post:
      tags:
        - 认证
      summary: 用户登出
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 登出成功
  /auth/refresh:
    post:
      tags:
        - 认证
      summary: 刷新Token
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 刷新成功
  /auth/info:
    get:
      tags:
        - 认证
      summary: 获取用户信息
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功
  /auth/login-history:
    get:
      tags:
        - 认证
      summary: 获取登录历史
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功

  # ==================== 用户接口 ====================
...
  /user/roles:
    get:
      tags:
        - 用户
      summary: 获取用户角色
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功

  # ==================== 密码重置接口 ====================
...
  /category/{id}:
    get:
      tags:
        - 分类
      summary: 获取分类详情
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 获取成功

    put:
      tags:
        - 分类
      summary: 更新分类
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                categoryName:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: 更新成功

    delete:
      tags:
        - 分类
      summary: 删除分类
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
  /category:
    post:
      tags:
        - 分类
      summary: 创建分类
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                categoryName:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: 创建成功

  # ==================== 标签接口 ====================
...
  /file/category/{id}:
    delete:
      tags:
        - 文件
      summary: 删除分类
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
...
  /analytics/pv:
    post:
      tags:
        - 数据分析
      summary: 记录页面访问
      description: 限流120次/分钟
      parameters:
        - name: pageUrl
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 记录成功
  /analytics/overview:
    get:
      tags:
        - 数据分析
      summary: 数据概览
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功
  /analytics/trend:
    get:
      tags:
        - 数据分析
      summary: 趋势数据
      security:
        - BearerAuth: []
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 获取成功
  /analytics/device:
    get:
      tags:
        - 数据分析
      summary: 设备统计
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功
  /analytics/source:
    get:
      tags:
        - 数据分析
      summary: 来源统计
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功
  /analytics/hot-articles:
    get:
      tags:
        - 数据分析
      summary: 热门文章
      security:
        - BearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
  /analytics/realtime:
    get:
      tags:
        - 数据分析
      summary: 实时数据
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功

  # ==================== 博客设置接口 ====================
...
  /admin/dashboard:
    get:
      tags:
        - 管理
      summary: 仪表盘数据
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功
  /admin/users:
    get:
      tags:
        - 管理
      summary: 用户列表
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 获取成功

    post:
      tags:
        - 管理
      summary: 创建用户
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                email:
                  type: string
      responses:
        '200':
          description: 创建成功
  /admin/users/{id}/status:
    put:
      tags:
        - 管理
      summary: 修改用户状态
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: status
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 修改成功
  /admin/users/{id}:
    delete:
      tags:
        - 管理
      summary: 删除用户
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
  /admin/users/{id}/roles:
    put:
      tags:
        - 管理
      summary: 分配角色
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
      responses:
        '200':
          description: 分配成功
  /admin/settings:
    get:
      tags:
        - 管理
      summary: 获取设置
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功
  /admin/settings/{key}:
    put:
      tags:
        - 管理
      summary: 更新设置
      security:
        - BearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
        - name: value
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 更新成功

  # ==================== 健康检查 ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Result:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
...